Extract Searchworks::Holdings to its own class
In Searchworks the bib data and holdings data come from the same request, but these are separate requests in Folio, so separating the two models now makes it easier to migrate
