+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2023-05-22 20:07:39 UTC using RuboCop version 1.50.2.
+# on 2023-05-25 21:50:50 UTC using RuboCop version 1.50.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -11,13 +11,12 @@ Capybara/SpecificActions:
   Exclude:
     - 'spec/features/mediation_table_spec.rb'
 
-# Offense count: 18
+# Offense count: 15
 Capybara/SpecificMatcher:
   Exclude:
     - 'spec/features/admin_requests_spec.rb'
     - 'spec/features/create_mediated_page_request_spec.rb'
     - 'spec/features/create_page_request_spec.rb'
-    - 'spec/features/item_selector_spec.rb'
     - 'spec/features/modal_layout_spec.rb'
     - 'spec/features/pickup_libraries_dropdown_spec.rb'
     - 'spec/features/send_request_buttons_spec.rb'
@@ -30,13 +29,13 @@ FactoryBot/SyntaxMethods:
   Exclude:
     - 'spec/models/request_spec.rb'
 
-# Offense count: 3
+# Offense count: 2
 # This cop supports safe autocorrection (--autocorrect).
 Layout/ClosingParenthesisIndentation:
   Exclude:
     - 'spec/features/admin_requests_spec.rb'
 
-# Offense count: 76
+# Offense count: 73
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: symmetrical, new_line, same_line
@@ -45,7 +44,7 @@ Layout/MultilineHashBraceLayout:
     - 'spec/factories/searchworks_api_json.rb'
     - 'spec/models/searchworks_item_spec.rb'
 
-# Offense count: 4
+# Offense count: 3
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: symmetrical, new_line, same_line
@@ -84,18 +83,18 @@ Lint/EmptyBlock:
     - 'spec/models/request_spec.rb'
     - 'spec/models/requests/page_spec.rb'
 
-# Offense count: 1
+# Offense count: 2
 Lint/MissingSuper:
   Exclude:
+    - 'app/components/library_instructions_component.rb'
     - 'app/lib/exceptions.rb'
-    - 'app/components/*'
 
 # Offense count: 1
 Lint/NoReturnInBeginEndBlocks:
   Exclude:
     - 'app/models/symphony/hold_record.rb'
 
-# Offense count: 6
+# Offense count: 5
 # Configuration parameters: CountComments, CountAsOne, AllowedMethods, AllowedPatterns.
 Metrics/MethodLength:
   Max: 11
@@ -139,7 +138,7 @@ Performance/RedundantEqualityComparisonBlock:
     - 'spec/models/paging_schedule_spec.rb'
     - 'spec/models/requests/mediated_page_spec.rb'
 
-# Offense count: 45
+# Offense count: 44
 RSpec/AnyInstance:
   Enabled: false
 
@@ -154,20 +153,35 @@ RSpec/BeforeAfterAll:
   Exclude:
     - 'spec/features/admin_requests_spec.rb'
 
-# Offense count: 81
+# Offense count: 80
 # Configuration parameters: Prefixes, AllowedPatterns.
 # Prefixes: when, with, without
 RSpec/ContextWording:
   Enabled: false
 
-# Offense count: 151
+# Offense count: 145
 # Configuration parameters: CountAsOne.
 RSpec/ExampleLength:
   Max: 46
 
-# Offense count: 42
+# Offense count: 40
 RSpec/ExpectInHook:
-  Enabled: false
+  Exclude:
+    - 'spec/controllers/mediated_pages_controller_spec.rb'
+    - 'spec/controllers/paging_schedule_controller_spec.rb'
+    - 'spec/features/mediation_table_spec.rb'
+    - 'spec/helpers/mediation_helper_spec.rb'
+    - 'spec/helpers/requests_helper_spec.rb'
+    - 'spec/jobs/expire_cdl_checkouts_job_spec.rb'
+    - 'spec/jobs/submit_borrow_direct_request_job_spec.rb'
+    - 'spec/jobs/submit_iplc_listener_job_spec.rb'
+    - 'spec/jobs/submit_reshare_request_job_spec.rb'
+    - 'spec/jobs/submit_symphony_request_job_spec.rb'
+    - 'spec/models/item_status_spec.rb'
+    - 'spec/models/requests/mediated_page_spec.rb'
+    - 'spec/models/requests/page_spec.rb'
+    - 'spec/requests/admin_comments_requests_spec.rb'
+    - 'spec/views/shared/_top_navbar.html.erb_spec.rb'
 
 # Offense count: 2
 # Configuration parameters: Include, CustomTransform, IgnoreMethods, SpecSuffixOnly.
@@ -189,17 +203,17 @@ RSpec/LetSetup:
   Exclude:
     - 'spec/features/mark_as_complete_spec.rb'
 
-# Offense count: 311
+# Offense count: 298
 RSpec/MultipleExpectations:
   Max: 16
 
-# Offense count: 472
+# Offense count: 431
 # Configuration parameters: EnforcedStyle, IgnoreSharedExamples.
 # SupportedStyles: always, named_only
 RSpec/NamedSubject:
   Enabled: false
 
-# Offense count: 107
+# Offense count: 97
 # Configuration parameters: AllowedGroups.
 RSpec/NestedGroups:
   Max: 5
@@ -267,11 +281,11 @@ RSpec/ScatteredSetup:
     - 'spec/controllers/scans_controller_spec.rb'
     - 'spec/features/mediation_table_spec.rb'
 
-# Offense count: 68
+# Offense count: 66
 RSpec/StubbedMock:
   Enabled: false
 
-# Offense count: 24
+# Offense count: 21
 RSpec/SubjectDeclaration:
   Exclude:
     - 'spec/helpers/requests_helper_spec.rb'
@@ -286,7 +300,7 @@ RSpec/SubjectDeclaration:
     - 'spec/models/searchworks_item_spec.rb'
     - 'spec/services/cdl_availability_spec.rb'
 
-# Offense count: 35
+# Offense count: 28
 RSpec/SubjectStub:
   Exclude:
     - 'spec/jobs/submit_borrow_direct_request_job_spec.rb'
@@ -298,7 +312,7 @@ RSpec/SubjectStub:
     - 'spec/models/symphony_response_spec.rb'
     - 'spec/models/user_spec.rb'
 
-# Offense count: 96
+# Offense count: 86
 # Configuration parameters: IgnoreNameless, IgnoreSymbolicNames.
 RSpec/VerifiedDoubles:
   Enabled: false
@@ -312,7 +326,7 @@ Rails/ApplicationRecord:
     - 'app/models/request.rb'
     - 'app/models/user.rb'
 
-# Offense count: 11
+# Offense count: 7
 # This cop supports unsafe autocorrection (--autocorrect-all).
 # Configuration parameters: NilOrEmpty, NotPresent, UnlessPresent.
 Rails/Blank:
@@ -386,13 +400,6 @@ Rails/OutputSafety:
     - 'app/controllers/application_controller.rb'
     - 'app/helpers/application_helper.rb'
 
-# Offense count: 1
-# Configuration parameters: Include.
-# Include: app/models/**/*.rb
-Rails/UniqueValidationWithoutIndex:
-  Exclude:
-    - 'app/models/user.rb'
-
 # Offense count: 6
 # This cop supports safe autocorrection (--autocorrect).
 Style/EmptyCaseCondition:
@@ -453,12 +460,13 @@ Style/NumericPredicate:
   Exclude:
     - 'app/models/dashboard.rb'
 
-# Offense count: 7
+# Offense count: 8
 Style/OpenStructUse:
   Exclude:
     - 'app/models/searchworks_item.rb'
     - 'spec/helpers/pickup_libraries_helper_spec.rb'
     - 'spec/models/request_spec.rb'
+    - 'spec/models/searchworks/holdings_spec.rb'
     - 'spec/models/searchworks_item_spec.rb'
     - 'spec/views/home/show.html.erb_spec.rb'
 
@@ -469,7 +477,7 @@ Style/OptionalBooleanParameter:
   Exclude:
     - 'app/models/searchworks_item.rb'
 
-# Offense count: 33
+# Offense count: 31
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: PreferredDelimiters.
 Style/PercentLiteralDelimiters:
@@ -496,11 +504,10 @@ Style/SelectByRegexp:
   Exclude:
     - 'app/models/location_rules.rb'
 
-# Offense count: 13
+# Offense count: 11
 # This cop supports unsafe autocorrection (--autocorrect-all).
 Style/SingleArgumentDig:
   Exclude:
-    - 'app/models/symphony/bib_info.rb'
     - 'app/models/symphony/catalog_info.rb'
     - 'app/models/symphony/circ_record.rb'
     - 'app/models/symphony/hold_record.rb'
@@ -529,7 +536,7 @@ Style/StringConcatenation:
 Style/SymbolArray:
   EnforcedStyle: brackets
 
-# Offense count: 4
+# Offense count: 7
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: WordRegex.
 # SupportedStyles: percent, brackets
