+++ app/models/searchworks_item.rb
@@ -42,10 +42,6 @@ class SearchworksItem
     @holdings ||= JSON.parse(json['holdings'].to_json, object_class: OpenStruct)
   end
 
-  def requested_holdings
-    @requested_holdings ||= RequestedHoldings.new(request, holdings)
-  end
-
   def finding_aid?
     !!json['finding_aid']
   end
@@ -81,61 +77,6 @@ class SearchworksItem
       JSON.parse(response.body)
     rescue JSON::ParserError
       {}
-    end
-  end
-
-  ###
-  #  ReqestedHoldings winnows down the entire
-  #  holdings to just what was requested by the user
-  ###
-  class RequestedHoldings
-    # @param [Request] request the users request
-    # @param [Array<#code>] holdings all of the holdings for the requested item
-    def initialize(request, holdings)
-      @request = request
-      @holdings = holdings
-    end
-
-    # @return [Array<OpenStruct>] a list of every holding in the requested library/location with the given barcodes
-    def where(barcodes: [])
-      return [] if barcodes.empty?
-
-      all.select do |item|
-        barcodes.include?(item.barcode)
-      end
-    end
-
-    # @return [Array<OpenStruct>] a list of every holding in the requested library/location
-    def all
-      return [] unless location.present?
-
-      location.items.map do |item|
-        item.request_status = @request.item_status(item.barcode)
-        item
-      end
-    end
-
-    def single_checked_out_item?
-      all.one? &&
-        all.first.current_location.try(:code) == 'CHECKEDOUT'
-    end
-
-    private
-
-    def library
-      return unless @holdings.present?
-
-      @holdings.find do |library|
-        library.code == @request.origin
-      end
-    end
-
-    def location
-      return unless library.present?
-
-      library.locations.find do |location|
-        location.code == @request.origin_location
-      end
     end
   end
 end
